refactor code to handle long literals; add implicit annotation for literal (not finished)

+++ src/intrange/qual/IntRange.java
@@ -11,7 +11,7 @@ import org.checkerframework.framework.qual.SubtypeOf;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE_PARAMETER, ElementType.TYPE_USE })
 public @interface IntRange {
-	int from() default Integer.MIN_VALUE;
-	int to() default Integer.MAX_VALUE;
+	long from() default Long.MIN_VALUE;
+	long to() default Long.MAX_VALUE;
 }
